- Use 'size_t' for memory sizes, positions and object identifiers (rather than 'unsigned int'). - Remove files related to automake/autoconf.

+++ sugs/sugs.h
@@ -58,7 +58,7 @@ class GSugs : public GALILEI::GPostCommunity
 
 	bool UseLevels;
 
-	unsigned int NbSugs;
+	size_t NbSugs;
 
 public:
